check for existance of cmake

+++ common/install_utils.sh
@@ -14,7 +14,11 @@ cmake_home=$(spack location -i cmake)
 export_cmake_path="export PATH=$(echo $cmake_home)/bin:$PATH"
 eval $export_cmake_path
 echo $export_cmake_path | tee -a /etc/profile
-ln -s $cmake_home/bin/cmake /bin/cmake
+
+if [ ! -e /bin/cmake ]
+then
+    ln -s $cmake_home/bin/cmake /bin/cmake
+fi
 
 # Install azcopy tool
 # To copy blobs or files to or from a storage account
